requirements.txt
locustio==0.14.5
boto3
aioboto3
